Move parameter types to metabase-lib
