+++ frontend/src/metabase-lib/lib/parameters/utils/operators.unit.spec.ts
@@ -1,4 +1,4 @@
-import { createMockUiParameter } from "metabase/parameters/mock";
+import { createMockUiParameter } from "metabase-lib/lib/parameters/mock";
 
 import {
   getOperatorDisplayName,
