+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebar.tsx
@@ -8,7 +8,6 @@ import Sidebar from "metabase/dashboard/components/Sidebar";
 
 import { isMappedExplicitActionButton } from "metabase/writeback/utils";
 
-import type { UiParameter } from "metabase/parameters/types";
 import type {
   Dashboard,
   DashboardOrderedCard,
@@ -19,6 +18,7 @@ import type {
   DatasetColumn,
 } from "metabase-types/api";
 import { isTableDisplay } from "metabase/lib/click-behavior";
+import type { UiParameter } from "metabase-lib/lib/parameters/types";
 import { clickBehaviorIsValid } from "metabase-lib/lib/parameters/utils/click-behavior";
 
 import { getColumnKey } from "metabase-lib/lib/queries/utils/get-column-key";
