+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/LinkOptions/LinkOptions.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback } from "react";
 import { t } from "ttag";
 
-import type { UiParameter } from "metabase/parameters/types";
 import type {
   DashboardOrderedCard,
   ArbitraryCustomDestinationClickBehavior,
@@ -10,6 +9,7 @@ import type {
   CustomDestinationClickBehaviorLinkType,
 } from "metabase-types/api";
 import { isTableDisplay } from "metabase/lib/click-behavior";
+import type { UiParameter } from "metabase-lib/lib/parameters/types";
 import { SidebarContent } from "../ClickBehaviorSidebar.styled";
 import CustomLinkText from "./CustomLinkText";
 import LinkedEntityPicker from "./LinkedEntityPicker";
