+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/TypeSelector/TypeSelector.tsx
@@ -5,8 +5,8 @@ import { color } from "metabase/lib/colors";
 
 import { isActionDashCard } from "metabase/writeback/utils";
 
-import type { UiParameter } from "metabase/parameters/types";
 import type { DashboardOrderedCard, ClickBehavior } from "metabase-types/api";
+import type { UiParameter } from "metabase-lib/lib/parameters/types";
 
 import { clickBehaviorOptions, getClickBehaviorOptionName } from "../utils";
 import { SidebarItem } from "../SidebarItem";
