+++ frontend/src/metabase/parameters/components/FormattedParameterValue/FormattedParameterValue.tsx
@@ -1,8 +1,11 @@
 import React from "react";
 
 import { formatParameterValue } from "metabase/parameters/utils/formatting";
-import { UiParameter, FieldFilterUiParameter } from "metabase/parameters/types";
 import ParameterFieldWidgetValue from "metabase/parameters/components/widgets/ParameterFieldWidget/ParameterFieldWidgetValue/ParameterFieldWidgetValue";
+import {
+  UiParameter,
+  FieldFilterUiParameter,
+} from "metabase-lib/lib/parameters/types";
 import { isDateParameter } from "metabase-lib/lib/parameters/utils/parameter-type";
 
 type FormattedParameterValueProps = {
