+++ frontend/src/metabase/parameters/utils/fields.ts
@@ -1,4 +1,4 @@
-import { UiParameter } from "metabase/parameters/types";
+import { UiParameter } from "metabase-lib/lib/parameters/types";
 
 export function hasFieldValues(parameter: UiParameter) {
   if ("fields" in parameter) {
