+++ frontend/src/metabase/parameters/utils/formatting.unit.spec.ts
@@ -1,6 +1,6 @@
-import { createMockUiParameter } from "metabase/parameters/mock";
-
 import { PRODUCTS, ORDERS } from "__support__/sample_database_fixture";
+import { createMockUiParameter } from "metabase-lib/lib/parameters/mock";
+
 import Field from "metabase-lib/lib/metadata/Field";
 import { formatParameterValue } from "./formatting";
