+++ frontend/src/metabase/parameters/utils/parameter-type.ts
@@ -1,6 +1,6 @@
 import _ from "underscore";
-import { FieldFilterUiParameter } from "metabase/parameters/types";
 import { Parameter } from "metabase-types/types/Parameter";
+import { FieldFilterUiParameter } from "metabase-lib/lib/parameters/types";
 import { FIELD_FILTER_PARAMETER_TYPES } from "metabase-lib/lib/parameters/constants";
 import { getParameterType } from "metabase-lib/lib/parameters/utils/parameter-type";
